requirements.txt
langgraph>=0.2.0
langchain-anthropic>=0.2.0
langchain-openai>=0.2.0
langchain-google-genai>=0.2.0
langchain-core>=0.3.0
pydantic>=2.0.0
streamlit>=1.39.0
plotly>=5.0.0
python-dotenv>=1.0.0
tenacity>=8.0.0